+++ server/src/dtsTypes/standardTypes/interrupts.ts
@@ -122,6 +122,20 @@ export default () => {
 
       const mapProperty = node.getProperty(`interrupt-map`);
       if (mapProperty) {
+        const addressCellsProperty = node.getProperty(`#address-cells`);
+        if (!addressCellsProperty) {
+          issues.push(
+            genStandardTypeDiagnostic(
+              StandardTypeIssue.PROPERTY_REQUIRES_OTHER_PROPERTY_IN_NODE,
+              property.ast,
+              DiagnosticSeverity.Error,
+              [...property.parent.nodeNameOrLabelRef],
+              [],
+              [property.name, "#address-cells", node.pathString]
+            )
+          );
+        }
+
         const match = parentInterruptNode.getNexusMapEntyMatch(
           "interrupt",
           macros,