+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- `reserved-memory` child node types.
+- Require of `#address-cells` to be defined when using `interrupt` property.
+- Require of `#address-cells` to be defined when using `interrupt-map` property.
+- Require of `#address-cells` to be defined when using `interrupts-extended` property.
+- Validation of `interrupt` property is linking to a `interrupt-map` entry.
+- Validation of `interrupts-extended` property is linking to a `interrupt-map` entry.
 - Validation of `interrupt-map-mask` has the correct number of `U32` values.
 - Validation of `interrupt-map` entries that none overlap.
 - Validation of `nexus-map-mask` has the correct number of `U32` values.